Fix: ユーザー認証情報取得機能をヘッダーに追加してしまっていたため、コールバック先のトップページに移管
